+++ src/Glpi/Console/Database/UpdateCommand.php
@@ -137,11 +137,6 @@ class UpdateCommand extends AbstractCommand implements ConfigurationCommandInter
 
     protected function initialize(InputInterface $input, OutputInterface $output)
     {
-
-        /** @var \Psr\SimpleCache\CacheInterface $GLPI_CACHE */
-        global $GLPI_CACHE;
-        $GLPI_CACHE = (new CacheManager())->getInstallerCacheInstance(); // Use dedicated "installer" cache
-
         parent::initialize($input, $output);
 
         $this->outputWarningOnMissingOptionnalRequirements();